pyproject.toml
[tool.poetry]
name = "cv-itmo"
version = "0.1.0"
description = "best map service"
authors = ["Kirill <gipokrat089@gmail.com>"]
readme = "README.md"
packages = [{include = "cv_itmo"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
sentinelsat = "^1.1.1"
rasterio = "^1.3.6"
matplotlib = "^3.7.0"
scikit-image = "^0.19.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
